requirements.txt
torch
numpy
pandas
matplotlib
seaborn
pyyaml
psutil
